refactor: move 'settlePendingOnChainPayment' into facade & improve function
